requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-default-group==1.2.4
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
feedparser==6.0.12
fsspec==2026.2.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
numpy==2.4.1
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
playwright==1.57.0
pluggy==1.6.0
postgrest==2.28.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
realtime==2.28.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.1
sqlite-fts4==1.0.3
sqlite-utils==3.39
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tabulate==0.9.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
supabase
feedparser
